+++ src/example.ts
@@ -1,5 +1,22 @@
 /**Basic Flutjs App With Tailwind */
-import {Spacer, Expanded,Row,RelativeUnits,BoxDecoration, Padding, Column, Container, Text, Colors, EdgeInsets, MainAxisAlignment, CrossAxisAlignment, TextStyle, Radius, AbsoluteUnits } from "./index.js";
+import {
+  Spacer,
+  Expanded,
+  Row,
+  RelativeUnits,
+  BoxDecoration,
+  Padding,
+  Column,
+  Container,
+  Text,
+  Colors,
+  EdgeInsets,
+  MainAxisAlignment,
+  CrossAxisAlignment,
+  TextStyle,
+  Radius,
+  AbsoluteUnits,
+} from "./index.js";
 import { BorderRadius } from "./painting/BorderRadiusGeometry.js";
 import { BoxShadow } from "./painting/boxShadow.js";
 import { State, StateManager } from "./state/stateManager.js";
@@ -71,6 +88,7 @@ WindiApp({
         }))
     }),
   }),
+
 });