+++ webpack.config.js
@@ -3,7 +3,7 @@ const path = require('path');
 module.exports = {
   entry: './dist/index.js',
   output: {
-    filename: 'flutjs.min.js',
+    filename: 'index.min.js',
     path: path.resolve(__dirname, 'dist'),
     
   },